refactor: move TokenRow to its own file
